requirements.txt
streamlit
pandas
python-telegram-bot